updated api insterest create to use datetime. migrated common stuff from generate-month to shared folders
